chore: Split functions crate into multi crates (#17133)
* chore: Split functions crate into multi crates

* split new crate databend-functions-scalar-arithmetic databend-functions-scalar-arithmetic-modulo databend-functions-scalar-datetime databend-functions-scalar-decimal-utils

* split numeric basic arithmetic and other bit arithmetic

* fix ut

* restore int/double arithmetic

* fix
