+++ src/query/functions/src/scalars/arithmetic/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "databend-functions-scalar-arithmetic"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+databend-common-expression = { workspace = true }
+databend-functions-scalar-decimal = { workspace = true }
+ethnum = { workspace = true }
+lexical-core = { workspace = true }
+match-template = { workspace = true }
+num-traits = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["match-template"]
