+++ src/query/functions/src/scalars/mathematics/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "databend-functions-scalar-math"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+crc32fast = { workspace = true }
+databend-common-base = { workspace = true }
+databend-common-expression = { workspace = true }
+databend-functions-scalar-decimal = { workspace = true }
+match-template = { workspace = true }
+num-traits = { workspace = true }
+
+[package.metadata.cargo-machete]
+ignored = ["match-template"]
