+++ src/query/functions/src/scalars/timestamp/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "databend-functions-scalar-datetime"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = { workspace = true }
+databend-common-column = { workspace = true }
+databend-common-exception = { workspace = true }
+databend-common-expression = { workspace = true }
+dtparse = { workspace = true }
+jiff = { workspace = true }
+num-traits = { workspace = true }
